requirements.txt
pytube>=15.0.0
PyYAML>=6.0.2